+++ tests/test_numerapi.py
@@ -94,3 +94,29 @@ def test_check_new_round(api):
         data = {"data": {"rounds": [{"openTime": open_time.isoformat()}]}}
         m.post(numerapi.numerapi.API_TOURNAMENT_URL, text=json.dumps(data))
         assert not api.check_new_round()
+
+
+def test_check_submission_successful(api):
+    test_cases = [
+        (True, None, False, True, 80, 0.69, False),
+        (False, False, False, True, 80, 0.69, False),
+        (False, True, True, None, 80, 0.69, False),
+        (False, True, False, False, 80, 0.69, False),
+        (False, True, False, True, 70, 0.69, False),
+        (False, True, False, True, 75, 0.70, False),
+        (False, True, False, True, 75, 0.69, True),
+    ]
+    with requests_mock.mock() as m:
+        for (originality_pending, originality_value, concordance_pending,
+             concordance_value, consistency, logloss, expected) in test_cases:
+            data = {"data": {"submissions": [
+              {"originality":
+                  {"pending": originality_pending, "value": originality_value},
+               "concordance":
+                  {"pending": concordance_pending, "value": concordance_value},
+               "consistency": consistency,
+               "validation_logloss": logloss}
+            ]}}
+        print(data)
+        m.post(numerapi.numerapi.API_TOURNAMENT_URL, text=json.dumps(data))
+        assert api.check_submission_successful(submission_id="foo") == expected